If available use zlib's Sync methods in benchmark.
