+++ package.json
@@ -41,6 +41,6 @@
     "multiparty": "^4.1.3",
     "ndoc": "*",
     "uglify-js": "*",
-    "zlibjs": "^0.2.0"
+    "zlibjs": "^0.3.1"
   }
 }
